+++ tests/columns/test_checkboxcolumn.py
@@ -18,3 +18,24 @@ def test_new_attrs_should_be_supported():
     assert attrs(table.rows[0]["col1"]) == {"type": "checkbox", "td_key": "td_value", "value": "data", "name": "col1"}
     assert attrs(table.columns["col2"].header) == {"type": "checkbox", "key": "value"}
     assert attrs(table.rows[0]["col2"]) == {"type": "checkbox", "key": "value", "value": "data", "name": "col2"}
+
+
+def column_is_checked():
+    class TestTable(tables.Table):
+        col = tables.CheckBoxColumn(attrs={"name": "col"}, checked='is_selected')
+
+    table = TestTable([{'col': '1', 'is_selected': True}, {'col': '2', 'is_selected': False}])
+    assert attrs(table.rows[0]["col"]) == {"type": "checkbox", "value": "1", "name": "col", "checked": "checked"}
+    assert attrs(table.rows[1]["col"]) == {"type": "checkbox", "value": "2", "name": "col"}
+
+
+def column_is_checked_callback():
+    def is_selected(value, record):
+        return value == '1'
+
+    class TestTable(tables.Table):
+        col = tables.CheckBoxColumn(attrs={"name": "col"}, checked=is_selected)
+
+    table = TestTable([{'col': '1'}, {'col': '2'}])
+    assert attrs(table.rows[0]["col"]) == {"type": "checkbox", "value": "1", "name": "col", "checked": "checked"}
+    assert attrs(table.rows[1]["col"]) == {"type": "checkbox", "value": "2", "name": "col"}